Sensor-flexible metadata parsing + compose the geometry plotter (#25)
Separate the sensor-specific work of discovering scene files and
extracting per-scene metadata from the sensor-agnostic stereo-pair
geometry math, and make StereoGeometryPlotter compose a parser instead
of inheriting from it.

- New asp_plot/sensors.py:
  - SensorMetadata ABC defining the reader interface (detect +
    get_scene_dicts) and the sensor-agnostic scene-dict schema.
  - DigitalGlobeMetadata(SensorMetadata): the existing DigitalGlobe/Maxar
    XML logic (file discovery, dg_mosaic tiling, per-scene extraction,
    ephemeris/attitude/footprint) moved verbatim out of the parser.
  - SENSORS registry + sensor_for_directory() detection, mirroring the
    bodies.py pattern, so adding ASTER/HiRISE/etc. is a new subclass with
    no changes to the geometry code.

- stereopair_metadata_parser.py is now a sensor-agnostic orchestrator:
  detects a reader, delegates scene extraction to it, and keeps only the
  pair-level geometry (convergence/BIE/asymmetry, intersection, bounds).

- stereo_geometry.py: StereoGeometryPlotter composes a
  StereopairMetadataParser via self.parser rather than subclassing it.

- Tests reorganized to mirror the split: new test_sensors.py (reader +
  detection), new test_stereopair_metadata_parser.py (pair geometry +
  formulas), and test_stereo_geometry.py trimmed to plotting +
  composition wiring. Imports updated.

Closes #25
